rbac-augment/backend/requirements.txt
# FastAPI核心依赖
fastapi==0.104.1
uvicorn[standard]==0.24.0

# 数据库相关
tortoise-orm==0.20.0
aerich==0.7.2
asyncpg==0.29.0  # PostgreSQL驱动
aiosqlite==0.19.0  # SQLite驱动

# 认证和安全
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# 工具库
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# HTTP客户端
httpx==0.25.2

# 日期时间处理
python-dateutil==2.8.2

# 开发工具
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0

# CORS支持
fastapi-cors==0.0.6
